fix: Rename lecture to resource
정신이 나갔는지 resource detail을 lecture detail이라 해서 수정했습니다.
