+++ src/containers/Shared/ModalProvider.js
@@ -1,0 +1,13 @@
+import React from "react";
+import ReportModalContainer from "./ModalContainers/ReportModalContainer";
+
+const ModalProvider = ({ children }) => {
+  return (
+    <>
+      {children}
+      <ReportModalContainer />
+    </>
+  );
+};
+
+export default ModalProvider;
